feat: useEvent hook
