+++ src/components/SearchBar/SearchBar.tsx
@@ -1,20 +1,11 @@
-import React, { useEffect, useState } from 'react';
-import { getDataFromLocalStorage, setDataToLocalStorage } from '../../domain/localStorageWorker';
+import { useEvent } from '../../hooks/useEvent';
+import React, { useState } from 'react';
+import { getDataFromLocalStorage } from '../../domain/localStorageWorker';
 import styles from './SearchBar.module.css';
 
 export const SearchBar = () => {
   const [searchValue, setSearchValue] = useState<string>(getDataFromLocalStorage() || '');
-
-  useEffect(() => {
-    const saveOnReload = () => {
-      setDataToLocalStorage(searchValue);
-    };
-    window.addEventListener('beforeunload', saveOnReload);
-    return () => {
-      setDataToLocalStorage(searchValue);
-      window.removeEventListener('beforeunload', saveOnReload);
-    };
-  }, [searchValue]);
+  useEvent(searchValue);
 
   const handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
     setSearchValue(e.target.value);
